+++ src/main/java/ssu/eatssu/domain/admin/controller/ManageFixMenuController.java
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
-import ssu.eatssu.domain.admin.dto.MenuBoards;
-import ssu.eatssu.domain.admin.dto.RegisterFixMenuRequest;
-import ssu.eatssu.domain.admin.dto.UpdateFixMenuRequest;
+import ssu.eatssu.domain.admin.dto.response.MenuBoards;
+import ssu.eatssu.domain.admin.dto.request.RegisterFixMenuRequest;
+import ssu.eatssu.domain.admin.dto.request.UpdateFixMenuRequest;
 import ssu.eatssu.domain.admin.service.ManageFixMenuService;
 import ssu.eatssu.domain.restaurant.entity.Restaurant;
 import ssu.eatssu.global.handler.response.BaseResponse;
