+++ src/main/java/ssu/eatssu/domain/admin/controller/ManageMealController.java
@@ -13,8 +13,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import ssu.eatssu.domain.admin.dto.MealInfo;
-import ssu.eatssu.domain.admin.dto.MenuBoards;
-import ssu.eatssu.domain.admin.dto.RegisterMealRequest;
+import ssu.eatssu.domain.admin.dto.response.MenuBoards;
+import ssu.eatssu.domain.admin.dto.request.RegisterMealRequest;
 import ssu.eatssu.domain.admin.service.ManageMealService;
 import ssu.eatssu.domain.menu.entity.constants.TimePart;
 import ssu.eatssu.domain.restaurant.entity.Restaurant;
