+++ src/main/java/ssu/eatssu/domain/admin/dto/SectionInBoard.java
@@ -1,5 +1,7 @@
 package ssu.eatssu.domain.admin.dto;
 
+import ssu.eatssu.domain.admin.dto.response.MenuLine;
+
 public interface SectionInBoard {
     void addMenuLine(MenuLine menuLine);
 }
