+++ src/main/java/ssu/eatssu/domain/admin/dto/request/RegisterCategoryRequest.java
@@ -1,4 +1,4 @@
-package ssu.eatssu.domain.admin.dto;
+package ssu.eatssu.domain.admin.dto.request;
 
 public record RegisterCategoryRequest(String name) {
 }
