+++ src/main/java/ssu/eatssu/domain/admin/dto/request/RegisterFixMenuRequest.java
@@ -1,4 +1,4 @@
-package ssu.eatssu.domain.admin.dto;
+package ssu.eatssu.domain.admin.dto.request;
 
 public record RegisterFixMenuRequest(Long categoryId, String name, Integer price) {
 }
