+++ src/main/java/ssu/eatssu/domain/admin/dto/request/RegisterMealRequest.java
@@ -1,4 +1,4 @@
-package ssu.eatssu.domain.admin.dto;
+package ssu.eatssu.domain.admin.dto.request;
 
 import java.util.List;
