+++ src/main/java/ssu/eatssu/domain/admin/dto/request/UpdateFixMenuRequest.java
@@ -1,4 +1,4 @@
-package ssu.eatssu.domain.admin.dto;
+package ssu.eatssu.domain.admin.dto.request;
 
 public record UpdateFixMenuRequest(String name, Integer price) {
 }
