+++ src/main/java/ssu/eatssu/domain/admin/dto/request/UpdateStatusRequest.java
@@ -1,4 +1,4 @@
-package ssu.eatssu.domain.admin.dto;
+package ssu.eatssu.domain.admin.dto.request;
 
 import ssu.eatssu.domain.inquiry.entity.InquiryStatus;
