+++ src/main/java/ssu/eatssu/domain/admin/dto/response/MenuBoards.java
@@ -1,4 +1,4 @@
-package ssu.eatssu.domain.admin.dto;
+package ssu.eatssu.domain.admin.dto.response;
 
 import java.util.ArrayList;
 import java.util.List;
