+++ src/main/java/ssu/eatssu/domain/admin/dto/response/MenuLine.java
@@ -1,4 +1,4 @@
-package ssu.eatssu.domain.admin.dto;
+package ssu.eatssu.domain.admin.dto.response;
 
 public record MenuLine(Long id, String name, Integer price, Double rating) {
 }
