+++ src/main/java/ssu/eatssu/domain/admin/service/ManageCategoryService.java
@@ -3,7 +3,7 @@ package ssu.eatssu.domain.admin.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ssu.eatssu.domain.admin.dto.RegisterCategoryRequest;
+import ssu.eatssu.domain.admin.dto.request.RegisterCategoryRequest;
 import ssu.eatssu.domain.admin.persistence.LoadCategoryRepository;
 import ssu.eatssu.domain.admin.persistence.ManageCategoryRepository;
 import ssu.eatssu.domain.menu.entity.MenuCategory;
