+++ src/main/java/ssu/eatssu/domain/admin/service/ManageFixMenuService.java
@@ -4,12 +4,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ssu.eatssu.domain.admin.dto.BriefMenu;
-import ssu.eatssu.domain.admin.dto.MenuBoard;
-import ssu.eatssu.domain.admin.dto.MenuBoards;
-import ssu.eatssu.domain.admin.dto.MenuLine;
-import ssu.eatssu.domain.admin.dto.MenuSection;
-import ssu.eatssu.domain.admin.dto.RegisterFixMenuRequest;
-import ssu.eatssu.domain.admin.dto.UpdateFixMenuRequest;
+import ssu.eatssu.domain.admin.dto.response.MenuBoard;
+import ssu.eatssu.domain.admin.dto.response.MenuBoards;
+import ssu.eatssu.domain.admin.dto.response.MenuLine;
+import ssu.eatssu.domain.admin.dto.response.MenuSection;
+import ssu.eatssu.domain.admin.dto.request.RegisterFixMenuRequest;
+import ssu.eatssu.domain.admin.dto.request.UpdateFixMenuRequest;
 import ssu.eatssu.domain.admin.persistence.LoadFixMenuRepository;
 import ssu.eatssu.domain.admin.persistence.ManageMenuRepository;
 import ssu.eatssu.domain.admin.persistence.MenuRatingRepository;
