+++ src/main/java/ssu/eatssu/domain/admin/service/ManageInquiryService.java
@@ -4,9 +4,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import ssu.eatssu.domain.admin.dto.InquiryLine;
-import ssu.eatssu.domain.admin.dto.PageWrapper;
-import ssu.eatssu.domain.admin.dto.UpdateStatusRequest;
+import ssu.eatssu.domain.admin.dto.response.InquiryLine;
+import ssu.eatssu.domain.admin.dto.response.PageWrapper;
+import ssu.eatssu.domain.admin.dto.request.UpdateStatusRequest;
 import ssu.eatssu.domain.admin.persistence.LoadInquiryRepository;
 import ssu.eatssu.domain.admin.persistence.ManageInquiryRepository;
 import ssu.eatssu.domain.inquiry.entity.Inquiry;
