+++ src/main/java/ssu/eatssu/domain/admin/service/ManageMealService.java
@@ -4,11 +4,11 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ssu.eatssu.domain.admin.dto.BriefMenu;
 import ssu.eatssu.domain.admin.dto.MealInfo;
-import ssu.eatssu.domain.admin.dto.MealSection;
-import ssu.eatssu.domain.admin.dto.MenuBoard;
-import ssu.eatssu.domain.admin.dto.MenuBoards;
-import ssu.eatssu.domain.admin.dto.MenuLine;
-import ssu.eatssu.domain.admin.dto.RegisterMealRequest;
+import ssu.eatssu.domain.admin.dto.response.MealSection;
+import ssu.eatssu.domain.admin.dto.response.MenuBoard;
+import ssu.eatssu.domain.admin.dto.response.MenuBoards;
+import ssu.eatssu.domain.admin.dto.response.MenuLine;
+import ssu.eatssu.domain.admin.dto.request.RegisterMealRequest;
 import ssu.eatssu.domain.admin.persistence.LoadMealRepository;
 import ssu.eatssu.domain.admin.persistence.ManageMealMenuRepository;
 import ssu.eatssu.domain.admin.persistence.ManageMealRepository;
